dls: move the most of all built-in operators from qualifier.c to the dsl layer
The built-in operators may be useful in the other engines in the future.
